+++ src/components/popover/index.js
@@ -2,8 +2,6 @@ import React, { useState, useEffect, useLayoutEffect, useRef, useContext, create
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { debounce } from 'lodash';
-import onClickOutside from 'react-onclickoutside';
-import { isFragment } from 'react-is';
 
 import TransparentButton from 'Components/buttons/transparent-button';
 import Button from 'Components/buttons/button';
@@ -17,57 +15,14 @@ popover pans, which have straight-walled sides rather than angled.
 ...also it's a [Bootstrap UI pattern](https://www.w3schools.com/bootstrap/bootstrap_popover.asp)
 */
 
-// statuses: 'closed' | 'openedFromKeyboard' | 'openedFromClick'
-const PopoverToggleContext = createContext('closed');
 
-const usePopoverToggle = ({ startOpen, onOpen }) => {
-  const [visible, setStatus] = useState(startOpen ? 'openedFromKeyboard' : 'closed');
-  const [openedFromKeyboard, setOpenedFromKeyboard] = useState(false);
-  
-  
-  
-  const setVisible = (newVisible) => {
-    if (!visible && newVisible && onOpen) {
-      onOpen();
-    }
-    setVisibleState(newVisible);
-  };
-
-  const togglePopover = (event) => {
-    setVisible(!visible);
-    if (event && event.detail === 0) {
-      // opened from keyboard
-      setOpenedFromKeyboard(true);
-    } else {
-      // opened from mouseclick
-      setOpenedFromKeyboard(false);
-    }
-  };
-
-  useEffect(() => {
-    if (!visible) return undefined;
-    const keyHandler = (event) => {
-      if (['Escape', 'Esc'].includes(event.key)) {
-        event.preventDefault();
-        setVisible(false);
-      }
-    };
-    window.addEventListener('keyup', keyHandler);
-    return () => window.removeEventListener('keyup', keyHandler);
-  }, [visible]);
-}
-
-
-
-
-const OpenedFromKeyboardContext = createContext(false);
 
 const useOpenedFromKeyboardFocus = (focusOnDialog) => {
   const ref = useRef();
-  const openedFromKeyboard = useContext(OpenedFromKeyboardContext);
+  const popoverStatus = useContext(PopoverToggleContext);
   useEffect(() => {
     const dialog = ref.current;
-    if (!dialog || !openedFromKeyboard) return;
+    if (!dialog || popoverStatus !== 'openedFromKeyboard') return;
     // focus on the dialog if there is only a destructible item in the popover that shouldn't automatically be focused on
     if (focusOnDialog) {
       dialog.focus();
@@ -79,7 +34,7 @@ const useOpenedFromKeyboardFocus = (focusOnDialog) => {
         focusableDialogElements[0].focus();
       }
     }
-  }, [openedFromKeyboard]);
+  }, [popoverStatus]);
   return ref;
 };
 
@@ -157,79 +112,7 @@ const styled = (Component, baseClassName) => ({ className, ...props }) => <Compo
 export const PopoverTitle = styled('div', styles.popoverTitle);
 export const InfoDescription = styled('p', styles.infoDescription);
 
-class UnmonitoredComponent extends React.Component {
-  handleClickOutside() {
-    this.props.onClickOutside();
-  }
 
-  render() {
-    return this.props.children;
-  }
-}
-
-const MonitoredComponent = onClickOutside(UnmonitoredComponent);
-
-
-
-export const PopoverContainer = ({ children, onOpen, outer, startOpen }) => {
-  const [visible, setVisibleState] = useState(startOpen);
-  const [openedFromKeyboard, setOpenedFromKeyboard] = useState(false);
-  const setVisible = (newVisible) => {
-    if (!visible && newVisible && onOpen) onOpen();
-    setVisibleState(newVisible);
-  };
-
-  const togglePopover = (event) => {
-    setVisible(!visible);
-    if (event && event.detail === 0) {
-      // opened from keyboard
-      setOpenedFromKeyboard(true);
-    } else {
-      // opened from mouseclick
-      setOpenedFromKeyboard(false);
-    }
-  };
-
-  useEffect(() => {
-    if (!visible) return undefined;
-    const keyHandler = (event) => {
-      if (['Escape', 'Esc'].includes(event.key)) {
-        event.preventDefault();
-        setVisible(false);
-      }
-    };
-    window.addEventListener('keyup', keyHandler);
-    return () => window.removeEventListener('keyup', keyHandler);
-  }, [visible]);
-
-  const props = { visible, setVisible, togglePopover };
-
-  const inner = children(props);
-  if (isFragment(inner)) {
-    console.error('PopoverContainer does not support Fragment as the top level item. Please use a different element.');
-  }
-  const before = outer ? outer(props) : null;
-
-  return (
-    <OpenedFromKeyboardContext.Provider value={openedFromKeyboard}>
-      {before}
-      <MonitoredComponent excludeScrollbar onClickOutside={() => setVisible(false)}>
-        {inner}
-      </MonitoredComponent>
-    </OpenedFromKeyboardContext.Provider>
-  );
-};
-PopoverContainer.propTypes = {
-  children: PropTypes.func.isRequired,
-  onOpen: PropTypes.func,
-  outer: PropTypes.func,
-  startOpen: PropTypes.bool,
-};
-PopoverContainer.defaultProps = {
-  onOpen: null,
-  outer: null,
-  startOpen: false,
-};
 
 const NestedPopoverContext = createContext();
